+++ mysql_rep_admin.py
@@ -445,7 +445,17 @@ def chk_slv_time(master, slaves, **kwargs):
 
     if frmt == "JSON":
         outdata = add_miss_slaves(master, outdata)
-        mongo_libs.json_prt_ins_2_db(outdata, **kwargs)
+        jdata = json.dumps(outdata, indent=4)
+        mongo_cfg = kwargs.get("class_cfg", None)
+        db_tbl = kwargs.get("db_tbl", None)
+        ofile = kwargs.get("ofile", None)
+
+        if mongo_cfg and db_tbl:
+            db, tbl = db_tbl.split(":")
+            mongo_libs.ins_doc(mongo_cfg, db, tbl, outdata)
+
+        if ofile:
+            gen_libs.write_file(ofile, "w", jdata)
 
 
 def chk_slv_other(master, slaves, **kwargs):